Move Sextuple.kt from model.viewmodel -> model package. Although it is only used in a ViewModel class currently, it was not meant to exclusively be used in a ViewModel class only.
